refactor($compile): 비속어 필터 외부 API로직 상품 컨텍스트로 이동
